These two ruby files are some code that I extended and improved.
Here are the extensions I made:
Add a require command to the tests so they can reference tree_node.
Add test coverage for methods: name and name=.
Add an instance method that returns true if the argument is its child.
Add an instance method that returns path to the calling node and write a test.
Write a method to help depth-first search. The method takes a block
argument and yields the node and it's descendants in order.