This is my code to solve a problem in Modern Compiler Implementation in ML.
The goal was to write code to interpret statements and expressions.
The statements and expressions can be nested to any level.

I implemented two mutually recursive functions, interpStm and interpExp.
interpStm receives a statement and a symbol table and returns a symbol table.
interpExp receives an integer and a symbol table and returns an integer and
a symbol table.

Statements cause side-effects on the variables in the symbol table or
something. Expressions evaluate to an integer. In either case, we must
return the symbol table, because the symbol table is not actually immutable.