This code solves a problem in Modern Compiler Implementation in ML.
The goal was to write code to interpret statements and expressions.
The statements and expressions can be nested to any level, but
nesting will limited by your machine's available resources.

I implemented two mutually recursive functions, interpStm and interpExp.
interpStm receives a statement and a symbol table and returns a symbol table.
interpExp receives an integer and a symbol table and returns an integer and
a symbol table.

Statements cause side-effects on the variables in the symbol table or print
something. Expressions evaluate to an integer. In either case, we must
return a symbol table, because the symbol table is actually immutable and
each returned symbol table is a new immutable value.